+++ src/main/kotlin/com/nylas/NylasClient.kt
@@ -402,6 +402,7 @@ class NylasClient(
               errorUri = "unknown",
               errorCode = "0",
               statusCode = response.code(),
+              headers = response.headers().toMultimap(),
             )
           }
 
@@ -423,6 +424,7 @@ class NylasClient(
               type = "unknown",
               message = "Unknown error received from the API: $responseBody",
               statusCode = response.code(),
+              headers = response.headers().toMultimap(),
             )
           }
           else -> throw ex
@@ -432,6 +434,7 @@ class NylasClient(
 
     if (parsedError != null) {
       parsedError.statusCode = response.code()
+      parsedError.headers = response.headers().toMultimap()
       throw parsedError
     }
 
@@ -439,6 +442,7 @@ class NylasClient(
       type = "unknown",
       message = "Unknown error received from the API: $responseBody",
       statusCode = response.code(),
+      headers = response.headers().toMultimap(),
     )
   }